feat: (config) Add support for custom configuration
Added support for custom configuration by using a JSON file located in the user's home directory. The configuration file contains information about the default input source, the list of apps to monitor and their corresponding input sources, and the list of apps to ignore. The configuration file can be modified by the user to suit their needs. When an app is activated, the script reads the configuration file to determine the corresponding input source and selects it automatically. Also added logging to better diagnose any issues that might arise.